Cypress: move verification tests to their own file
