+++ cypress/e2e/crypto/crypto.spec.ts
@@ -19,13 +19,7 @@ import type { VerificationRequest } from "matrix-js-sdk/src/crypto-api";
 import type { CypressBot } from "../../support/bot";
 import { HomeserverInstance } from "../../plugins/utils/homeserver";
 import { UserCredentials } from "../../support/login";
-import {
-    checkDeviceIsCrossSigned,
-    EmojiMapping,
-    handleVerificationRequest,
-    logIntoElement,
-    waitForVerificationRequest,
-} from "./utils";
+import { doTwoWaySasVerification, waitForVerificationRequest } from "./utils";
 import { skipIfRustCrypto } from "../../support/util";
 
 interface CryptoTestContext extends Mocha.Context {
@@ -106,27 +100,6 @@ function autoJoin(client: MatrixClient) {
             if (member.membership === "invite" && member.userId === client.getUserId()) {
                 client.joinRoom(member.roomId);
             }
-        });
-    });
-}
-
-/**
- * Given a VerificationRequest in a bot client, add cypress commands to:
- *   - wait for the bot to receive a 'verify by emoji' notification
- *   - check that the bot sees the same emoji as the application
- *
- * @param botVerificationRequest - a verification request in a bot client
- */
-function doTwoWaySasVerification(botVerificationRequest: VerificationRequest): void {
-    // on the bot side, wait for the emojis, confirm they match, and return them
-    const emojiPromise = handleVerificationRequest(botVerificationRequest);
-
-    // then, check that our application shows an emoji panel with the same emojis.
-    cy.wrap(emojiPromise).then((emojis: EmojiMapping[]) => {
-        cy.get(".mx_VerificationShowSas_emojiSas_block").then((emojiBlocks) => {
-            emojis.forEach((emoji: EmojiMapping, index: number) => {
-                expect(emojiBlocks[index].textContent.toLowerCase()).to.eq(emoji[0] + emoji[1]);
-            });
         });
     });
 }
@@ -406,70 +379,5 @@ describe("Cryptography", function () {
                 .should("have.class", "mx_EventTile_verified")
                 .should("not.have.descendants", ".mx_EventTile_e2eIcon_warning");
         });
-    });
-});
-
-describe("Verify own device", () => {
-    let aliceBotClient: CypressBot;
-    let homeserver: HomeserverInstance;
-
-    beforeEach(() => {
-        skipIfRustCrypto();
-        cy.startHomeserver("default").then((data: HomeserverInstance) => {
-            homeserver = data;
-
-            // Visit the login page of the app, to load the matrix sdk
-            cy.visit("/#/login");
-
-            // wait for the page to load
-            cy.window({ log: false }).should("have.property", "matrixcs");
-
-            // Create a new device for alice
-            cy.getBot(homeserver, { bootstrapCrossSigning: true }).then((bot) => {
-                aliceBotClient = bot;
-            });
-        });
-    });
-
-    afterEach(() => {
-        cy.stopHomeserver(homeserver);
-    });
-
-    /* Click the "Verify with another device" button, and have the bot client auto-accept it.
-     *
-     * Stores the incoming `VerificationRequest` on the bot client as `@verificationRequest`.
-     */
-    function initiateAliceVerificationRequest() {
-        // alice bot waits for verification request
-        const promiseVerificationRequest = waitForVerificationRequest(aliceBotClient);
-
-        // Click on "Verify with another device"
-        cy.get(".mx_AuthPage").within(() => {
-            cy.findByRole("button", { name: "Verify with another device" }).click();
-        });
-
-        // alice bot responds yes to verification request from alice
-        cy.wrap(promiseVerificationRequest).as("verificationRequest");
-    }
-
-    it("with SAS", function (this: CryptoTestContext) {
-        logIntoElement(homeserver.baseUrl, aliceBotClient.getUserId(), aliceBotClient.__cypress_password);
-
-        // Launch the verification request between alice and the bot
-        initiateAliceVerificationRequest();
-
-        // Handle emoji SAS verification
-        cy.get(".mx_InfoDialog").within(() => {
-            cy.get<VerificationRequest>("@verificationRequest").then((request: VerificationRequest) => {
-                // Handle emoji request and check that emojis are matching
-                doTwoWaySasVerification(request);
-            });
-
-            cy.findByRole("button", { name: "They match" }).click();
-            cy.findByRole("button", { name: "Got it" }).click();
-        });
-
-        // Check that our device is now cross-signed
-        checkDeviceIsCrossSigned();
     });
 });
